se hacen cambios

+++ programacion_web/ejemplos/PHP/3-PHP_avanzado/7-conexionBDMySQLPDO/app/controllers/IngresosController.php
@@ -1,35 +1,41 @@
-<?php 
+<?php
+
 /**
-* Los 7 métodos que suelen tener los controladores:
-* index: muestra la lista de todos los recursos.
-* create: muestra un formulario para ingresar un nuevo recurso. (luego manda a llamar al método store).
-* store: registra dentro de la base de datos el nuevo recurso.
-* show: muestra un recurso específico.
-* edit: muestra un formulario para editar un recurso. (luego manda a llamar al método update).
-* update: actualiza el recurso dentro de la base de datos.
-* destroy: elimina un recurso.
-*/
+ * Los 7 métodos que suelen tener los controladores:
+ * index: muestra la lista de todos los recursos.
+ * create: muestra un formulario para ingresar un nuevo recurso. (luego manda a llamar al método store).
+ * store: registra dentro de la base de datos el nuevo recurso.
+ * show: muestra un recurso específico.
+ * edit: muestra un formulario para editar un recurso. (luego manda a llamar al método update).
+ * update: actualiza el recurso dentro de la base de datos.
+ * destroy: elimina un recurso.
+ */
 require_once("../app/models/ConexionModel.php");
 
-class IngresosController {
+class IngresosController
+{
     private $conn;
 
-    public function __construct(){
+    public function __construct()
+    {
         $this->conn = ConexionModel::getInstance()->getDatabaseInstance();
     }
 
-    public function index(){
+    public function index()
+    {
         $consulta = $this->conn->prepare("SELECT * from ingresos;");
         $consulta->execute();
         $resultados = $consulta->fetchAll();
         require_once("../app/views/ingresos/index.php");
     }
 
-    public function create(){
+    public function create()
+    {
         require_once("../app/views/ingresos/create.php");
     }
 
-    public function store($data){
+    public function store($data)
+    {
         $consulta = $this->conn->prepare("INSERT INTO ingresos(metodo_pago,tipo,fecha_ingreso,cantidad,descripcion) VALUES (:metodo_pago,:tipo,:fecha_ingreso,:cantidad,:descripcion);");
         $consulta->bindValue(":metodo_pago", $data['metodo_pago']);
         $consulta->bindValue(":tipo", $data['tipo']);
@@ -39,8 +45,9 @@ class IngresosController {
         $consulta->execute();
         header("location: ingresos");
     }
-    
-    public function show($id){
+
+    public function show($id)
+    {
         $consulta = $this->conn->prepare("SELECT * FROM ingresos WHERE id=:id;");
         $consulta->execute([
             ":id" => $id
@@ -49,9 +56,12 @@ class IngresosController {
         require_once("../app/views/ingresos/ingreso.php");
     }
 
-    public function edit(){}
+    public function edit()
+    {
+    }
 
-    public function update($data, $id){
+    public function update($data, $id)
+    {
         $consulta = $this->conn->prepare("UPDATE ingresos SET 
         metodo_pago = :metodo_pago,
         tipo = :tipo,
@@ -65,24 +75,19 @@ class IngresosController {
             ":fecha_ingreso" => $data['fecha_ingreso'],
             ":cantidad" => $data['cantidad'],
             ":descripcion" => $data['descripcion'],
-        ]);         
+        ]);
     }
 
-    public function destroy($id){
-        $consulta = $this->conn->beginTransaction();
-        $consulta = $this->conn->prepare("DELETE FROM ingresos WHERE id=:id;");
-        $consulta->execute([
-            ":id" => $id
-        ]); 
-        $validar = readline("Esta completamente seguro que desea eliminar el registro?");
-        if($validar == 'no') {
-            $this->conn->rollBack();
-        } elseif($validar == 'si') {
-            $this->conn->commit();
+    public function destroy($id)
+    {
+        if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+            $consulta = $this->conn->prepare("DELETE FROM ingresos WHERE id=:id;");
+            $consulta->execute([
+                ":id" => $id
+            ]);
+            header("location: ingresos");
         } else {
-            echo "Error al procesar la transacción";
+            echo "Método no permitido";
         }
     }
 }
-
-?>
